Исправлены ошибки

+++ Исходные файлы/Bouble_Sort.cpp
@@ -3,7 +3,7 @@
 
 
 
-void Bouble_Sort(int64_t* arr, int32_t size) {
+void Bouble_Sort(int32_t* arr, int32_t size) {
     for (size_t i = 0; i < size; ++i) {
         bool flag = true;
         for (int j = 0; j < size - (i + 1); j++) {
