+++ Файлы заголовков/rm_minus_num.h
@@ -5,7 +5,7 @@
 
 #include <iostream>
 #include <algorithm>
-void Negative_Numbers_Are_Equal_To_Zero(int64_t*, int32_t);
+void Negative_Numbers_Are_Equal_To_Zero(int32_t*, int32_t);
 
 
 #endif
